Phase 3/developed_version/requirements.txt
Flask
Flask-CORS
Flask-Session
pandas
python-dotenv
requests
numpy
setuptools
google.generativeai
